Reworked on fluent api. Created LinqLikeBuilder to start sql quuery directly with Select clause
