+++ RawSqlHelper.LinqLikeExtension/LinqLikeExtension.cs
@@ -8,27 +8,6 @@
         public const string CommaSeparator = ", ";
 
         /// <summary>
-        /// Adds 'SELECT' statement. If <paramref name="fields"/> is <see langword="null"/> or empty, then '*' value used
-        /// </summary>
-        /// <param name="builder"></param>
-        /// <param name="fields">Fields to select.</param>
-        /// <returns></returns>
-        public static SqlQueryBuilder Select(this SqlQueryBuilder builder, params string[] fields)
-        {
-            if (fields == null || fields.Length == 0)
-            {
-                fields = new string[] { };
-            }
-            var parameters = string.Join(CommaSeparator, fields);
-            if (string.IsNullOrEmpty(parameters))
-            {
-                parameters = "*";
-            }
-            var select = $"{SelectKey} {parameters}";
-            return builder.Add(select);
-        }
-
-        /// <summary>
         /// Adds 'FROM' statement.
         /// </summary>
         /// <param name="builder"></param>
@@ -42,6 +21,23 @@
             }
 
             var from = $"{FromKey} {tableName}";
+            return builder.Add(from);
+        }
+
+        /// <summary>
+        /// Adds 'FROM' statement.
+        /// </summary>
+        /// <param name="builder"></param>
+        /// <param name="subselect">Subselect for 'FROM' clause</param>
+        /// <returns></returns>
+        public static SqlQueryBuilder FromSubselect(this SqlQueryBuilder builder, string subselect)
+        {
+            if (string.IsNullOrEmpty(subselect))
+            {
+                throw new System.ArgumentNullException(nameof(subselect));
+            }
+
+            var from = $"{FromKey} {subselect.Brackets()}";
             return builder.Add(from);
         }
